requirements.txt
numpy~=1.21.2
uvicorn~=0.24.0.post1
fastapi~=0.104.1
Pillow~=8.4.0
opencv-python~=4.5.5.62
./wheel/arena_api-2.3.3-py3-none-any.whl
requests~=2.31.0